+++ ThirdPartyAdapters/facebook/facebook/src/main/java/com/google/ads/mediation/facebook/FacebookInitializer.java
@@ -3,9 +3,9 @@ package com.google.ads.mediation.facebook;
 import android.content.Context;
 
 import com.facebook.ads.AudienceNetworkAds;
-import com.google.android.gms.ads.MobileAds;
 
 import java.util.ArrayList;
+
 
 class FacebookInitializer implements AudienceNetworkAds.InitListener {
 
@@ -48,8 +48,7 @@ class FacebookInitializer implements AudienceNetworkAds.InitListener {
 
         getInstance().mListeners.add(listener);
         AudienceNetworkAds.buildInitSettings(context)
-                .withMediationService("GOOGLE_" + MobileAds.getVersionString()
-                    + ":" + BuildConfig.VERSION_NAME)
+                .withMediationService("GOOGLE:"+ BuildConfig.VERSION_NAME)
                 .withPlacementIds(placements)
                 .withInitListener(FacebookInitializer.this)
                 .initialize();