+++ ThirdPartyAdapters/ironsource/ironsource/src/main/java/com/google/ads/mediation/ironsource/IronSourceAdapter.java
@@ -10,17 +10,26 @@ import com.google.android.gms.ads.AdRequest;
 import com.google.android.gms.ads.mediation.MediationAdRequest;
 import com.google.android.gms.ads.mediation.MediationInterstitialAdapter;
 import com.google.android.gms.ads.mediation.MediationInterstitialListener;
+import com.google.ads.mediation.ironsource.IronSourceMediationAdapter.INSTANCE_STATE;
 
 import com.ironsource.mediationsdk.IronSource;
 import com.ironsource.mediationsdk.logger.IronSourceError;
-import com.ironsource.mediationsdk.sdk.ISDemandOnlyInterstitialListener;
+
+import java.lang.ref.WeakReference;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
+import java.util.concurrent.atomic.AtomicBoolean;
+
+import static com.google.ads.mediation.ironsource.IronSourceAdapterUtils.DEFAULT_INSTANCE_ID;
+import static com.google.ads.mediation.ironsource.IronSourceAdapterUtils.TAG;
 
 /**
  * A {@link MediationInterstitialAdapter} to load and show IronSource interstitial ads using Google
  * Mobile Ads SDK mediation.
  */
 public class IronSourceAdapter
-        implements MediationInterstitialAdapter, ISDemandOnlyInterstitialListener {
+        implements MediationInterstitialAdapter {
 
     /**
      * Mediation interstitial ad listener used to forward interstitial events from
@@ -32,6 +41,27 @@ public class IronSourceAdapter
      * This is the id of the instance to be shown.
      */
     private String mInstanceID;
+
+    private static AtomicBoolean mDidInitInterstitial = new AtomicBoolean(false);
+
+    private final static List<IronSource.AD_UNIT> mAdUnitsToInit =
+            new ArrayList<>(Collections.singletonList(IronSource.AD_UNIT.INTERSTITIAL));
+
+    /**
+     * Holds the interstitial instance state
+     */
+    private INSTANCE_STATE mState = INSTANCE_STATE.START;
+
+    /**
+     * This is the id of interstitial instance requested.
+     */
+    INSTANCE_STATE getInstanceState() {
+        return mState;
+    }
+
+    void setInstanceState(IronSourceMediationAdapter.INSTANCE_STATE mState) {
+        this.mState = mState;
+    }
 
     //region MediationInterstitialAdapter implementation.
     @Override
@@ -46,50 +76,42 @@ public class IronSourceAdapter
         if (!(context instanceof Activity)) {
             // Context not an Activity context, log the reason for failure and fail the
             // initialization.
-            Log.e(IronSourceAdapterUtils.TAG, "IronSource SDK requires " +
-                    "an Activity context to initialize");
+            Log.e(TAG, "IronSource SDK requires an Activity context to initialize");
             onISAdFailedToLoad(AdRequest.ERROR_CODE_INVALID_REQUEST);
             return;
         }
 
-        try {
-            String appKey = serverParameters.getString(IronSourceAdapterUtils.KEY_APP_KEY);
-            if (TextUtils.isEmpty(appKey)) {
-                Log.w(IronSourceAdapterUtils.TAG,
-                        "Initialization Failed: Missing or Invalid App Key.");
-                onISAdFailedToLoad(AdRequest.ERROR_CODE_INVALID_REQUEST);
-                return;
+        String appKey = serverParameters.getString(IronSourceAdapterUtils.KEY_APP_KEY);
+        if (TextUtils.isEmpty(appKey)) {
+            Log.e(TAG, String.format("IronSource initialization failed," +
+                    "make sure that the '%s' server parameter is added"
+                    , IronSourceAdapterUtils.KEY_APP_KEY));
+            onISAdFailedToLoad(AdRequest.ERROR_CODE_INVALID_REQUEST);
+            return;
+        }
+
+        this.mInstanceID = serverParameters.getString(IronSourceAdapterUtils.KEY_INSTANCE_ID, DEFAULT_INSTANCE_ID);
+        if (!mDidInitInterstitial.getAndSet(true)) {
+
+            try {
+                Log.d(TAG, String.format("Init IronSource interstitial ad for instance: %s", this.mInstanceID));
+                IronSourceManager.getInstance().initIronSourceSDK((Activity) context, appKey, mAdUnitsToInit);
+                Log.d(TAG, String.format("IronSource load Interstitial called with instanceId: %s", this.mInstanceID));
+                IronSourceManager.getInstance().loadInterstitial(this.mInstanceID, new WeakReference<>(IronSourceAdapter.this));
+            } catch (Exception e) {
+                Log.e(TAG, String.format("IronSource initialization failed with Error: %s", e.getMessage()));
+                onISAdFailedToLoad(AdRequest.ERROR_CODE_INTERNAL_ERROR);
             }
-
-            this.mInstanceID = serverParameters.getString(
-                    IronSourceAdapterUtils.KEY_INSTANCE_ID, "0");
-
-            IronSource.setISDemandOnlyInterstitialListener(this);
-            if (!IronSourceAdapterUtils.isIronSourceInitialized(IronSource.AD_UNIT.INTERSTITIAL)) {
-                IronSourceAdapterUtils.initIronSourceSDK((Activity) context, appKey,
-                        IronSource.AD_UNIT.INTERSTITIAL);
-            }
-
-            Log.d(IronSourceAdapterUtils.TAG, "Load IronSource interstitial ad for instance: " +
-                    this.mInstanceID);
-            IronSource.loadISDemandOnlyInterstitial(this.mInstanceID);
-        } catch (Exception e) {
-            Log.w(IronSourceAdapterUtils.TAG, "Initialization Failed.", e);
-            onISAdFailedToLoad(AdRequest.ERROR_CODE_INTERNAL_ERROR);
+        } else {
+            Log.d(TAG, String.format("IronSource loadInterstitial called with instanceId: %s", this.mInstanceID));
+            IronSourceManager.getInstance().loadInterstitial(mInstanceID, new WeakReference<>(IronSourceAdapter.this));
         }
     }
 
     @Override
     public void showInterstitial() {
-        try {
-            IronSource.showISDemandOnlyInterstitial(this.mInstanceID);
-        } catch (Exception e) {
-            Log.w(IronSourceAdapterUtils.TAG, "IronSource Interstitial failed to show.", e);
-            if (mInterstitialListener != null) {
-                mInterstitialListener.onAdOpened(IronSourceAdapter.this);
-                mInterstitialListener.onAdClosed(IronSourceAdapter.this);
-            }
-        }
+        Log.d(TAG, String.format("IronSource showInterstitial for instance: %s", this.mInstanceID));
+        IronSourceManager.getInstance().showInterstitial(mInstanceID);
     }
     //endregion
 
@@ -109,8 +131,8 @@ public class IronSourceAdapter
      * Private IronSource methods
      */
     private void onISAdFailedToLoad(final int errorCode) {
-        Log.e(IronSourceAdapterUtils.TAG, String.format("IronSource Interstitial failed to load " +
-                "for instance %s, Error: %d", this.mInstanceID, errorCode));
+        Log.e(TAG, String.format("IronSource Interstitial failed to load for instance %s, Error: %d"
+                , this.mInstanceID, errorCode));
 
         if (mInterstitialListener != null) {
             IronSourceAdapterUtils.sendEventOnUIThread(new Runnable() {
@@ -122,14 +144,9 @@ public class IronSourceAdapter
     }
 
     //region ISDemandOnlyInterstitialListener implementation.
-    @Override
     public void onInterstitialAdReady(String instanceId) {
-        Log.d(IronSourceAdapterUtils.TAG, "IronSource Interstitial loaded successfully " +
-                "for instance " + instanceId);
-
-        // We only listen to a registered instance.
-        if (!this.mInstanceID.equals(instanceId))
-            return;
+        Log.d(TAG, String.format("IronSource Interstitial loaded successfully for instance %s "
+                , instanceId));
 
         if (mInterstitialListener != null) {
             IronSourceAdapterUtils.sendEventOnUIThread(new Runnable() {
@@ -140,10 +157,9 @@ public class IronSourceAdapter
         }
     }
 
-    @Override
     public void onInterstitialAdLoadFailed(String instanceId, IronSourceError ironSourceError) {
-        Log.e(IronSourceAdapterUtils.TAG, String.format("IronSource Interstitial failed to load " +
-                "for instance %s, Error: %s", this.mInstanceID, ironSourceError.getErrorMessage()));
+        Log.e(TAG, String.format("IronSource Interstitial failed to load for instance %s  with Error: %s"
+                , instanceId, ironSourceError.getErrorMessage()));
 
         // We only listen to a registered instance.
         if (!this.mInstanceID.equals(instanceId))
@@ -152,10 +168,9 @@ public class IronSourceAdapter
         onISAdFailedToLoad(AdRequest.ERROR_CODE_INTERNAL_ERROR);
     }
 
-    @Override
     public void onInterstitialAdOpened(String instanceId) {
-        Log.d(IronSourceAdapterUtils.TAG, "IronSource Interstitial opened ad for instance " +
-                instanceId);
+        Log.d(TAG, String.format("IronSource Interstitial opened ad for instance %s",
+                instanceId));
 
         if (mInterstitialListener != null) {
             IronSourceAdapterUtils.sendEventOnUIThread(new Runnable() {
@@ -166,10 +181,9 @@ public class IronSourceAdapter
         }
     }
 
-    @Override
     public void onInterstitialAdClosed(String instanceId) {
-        Log.d(IronSourceAdapterUtils.TAG, "IronSource Interstitial closed ad for instance " +
-                instanceId);
+        Log.d(TAG, String.format("IronSource Interstitial closed ad for instance %s",
+                instanceId));
 
         if (mInterstitialListener != null) {
             IronSourceAdapterUtils.sendEventOnUIThread(new Runnable() {
@@ -180,15 +194,9 @@ public class IronSourceAdapter
         }
     }
 
-    @Override
-    public void onInterstitialAdShowSucceeded(String instanceId) {
-        // No relevant delegate in AdMob interface.
-    }
-
-    @Override
     public void onInterstitialAdShowFailed(String instanceId, IronSourceError ironSourceError) {
-        Log.e(IronSourceAdapterUtils.TAG, String.format("IronSource Interstitial failed to show " +
-                "for instance %s, Error: %s", this.mInstanceID, ironSourceError.getErrorMessage()));
+        Log.e(TAG, String.format("IronSource Interstitial failed to show " +
+                "for instance %s with Error: %s", instanceId, ironSourceError.getErrorMessage()));
         if (mInterstitialListener != null) {
             IronSourceAdapterUtils.sendEventOnUIThread(new Runnable() {
                 @Override
@@ -200,10 +208,9 @@ public class IronSourceAdapter
         }
     }
 
-    @Override
     public void onInterstitialAdClicked(String instanceId) {
-        Log.d(IronSourceAdapterUtils.TAG, "IronSource Interstitial clicked ad for instance " +
-                instanceId);
+        Log.d(TAG, String.format("IronSource Interstitial ad clicked for instance %s",
+                instanceId));
 
         if (mInterstitialListener != null) {
             IronSourceAdapterUtils.sendEventOnUIThread(new Runnable() {